requirements.txt
pyautogui
opencv-python
numpy
pillow